pyproject.toml
[project]
name = "api-anoninbox"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (==5.2)",
    "drf-spectacular (>=0.29.0,<0.30.0)",
    "drf-spectacular[swagger] (>=0.29.0,<0.30.0)",
    "drf-spectacular[redoc] (>=0.29.0,<0.30.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)"
]

[dependency-groups]
dev = [
    "djangorestframework (==3.16.1)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{include = "anoninbox"}]
